apps/agents/scripts/apply_migration_004.py
import os
from supabase import create_client, Client
from dotenv import load_dotenv
from pathlib import Path

# Setup Environment
script_dir = Path(__file__).resolve().parent
project_root = script_dir.parents[2]
env_path = project_root / ".env.local"

if env_path.exists():
    load_dotenv(env_path)
else:
    load_dotenv(project_root / ".env")

url = os.environ.get("NEXT_PUBLIC_SUPABASE_URL")
key = os.environ.get("SUPABASE_SERVICE_ROLE_KEY")

if not url or not key:
    raise ValueError("Missing Supabase credentials")

supabase: Client = create_client(url, key)

SQL_FILE = project_root / "packages/database/migrations/004_participant_selections.sql"

def apply_migration():
    print(f"Applying migration from {SQL_FILE}...")
    with open(SQL_FILE, 'r') as f:
        sql = f.read()

    # Split functionality might be needed if the client doesn't support multiple statements at once, 
    # but supabase-py 'rpc' or direct sql execution via some clients can be tricky.
    # However, supabase-py relies on PostgREST which doesn't allow raw SQL usually.
    # WE MUST CHECK if we have a way to run SQL.
    # Usually we don't via the JS/Python client unless there is an RPC for it or we use psycopg2.
    
    # Wait, previous tests used `supabase.table().insert()`. 
    # If I can't run DDL via the client, I should notify the user.
    # BUT, I see `pgcrypto` usage in 001_initial.sql.
    
    # Let's try to find an `exec_sql` function or similar if created in previous turns?
    # No, I don't see one.
    
    # Alternative: The user asked me to create the migration. I should just create the file and the SEED script.
    # The seed script will FAIL if the table doesn't exist.
    # So I must assume the user might need to run it, OR I use a workaround.
    
    # Workaround: If I can't run SQL, I will inform the user.
    # BUT, I'll try to just CREATE the tables using a direct psycopg2 connection if I can? 
    # No, I don't have the connection string (DB URI), only the API URL/Key.
    
    # NOTE: I will skip applying it via Python since I likely lack permissions/tools via the API client.
    # I will create the seed script. The User might have an auto-migration runner.
    # Actually, I'll `notify_user` telling them to apply it.
    pass

if __name__ == "__main__":
    print("Manual migration application required via SQL Editor or CLI.")
