apps/agents/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
python-multipart
aiofiles
supabase
redis
Faker
httpx
